socialNetworkML/src/MediaFilterService/requirements.txt
numpy==1.22.1
#numpy==1.19.0
pillow==8.2.0
h5py==2.10.0
#h5py==3.1.0
intel-tensorflow==2.9.0
#tensorflow-estimator==2.6.0
#tensorflow-cpu==2.1.0
tensorflow-hub==0.7.0
# --find-links https://download.pytorch.org/whl/torch_stable.html
# torch==1.9.0+cpu
# torchvision==0.10.0+cpu
